+++ keyboards/kyria/keymaps/ninjonas/rules.mk
@@ -2,3 +2,6 @@ OLED_DRIVER_ENABLE = yes   # Enables the use of OLED displays
 ENCODER_ENABLE = yes       # Enables the use of one or more encoders
 RGBLIGHT_ENABLE = yes      # Enable keyboard RGB underglow
 LINK_TIME_OPTIMIZATION_ENABLE = yes
+
+SRC += encoder.c \
+       oled.c
